chore: extract exercise filters to a hook
