+++ frontend/pages/index.tsx
@@ -1,18 +1,20 @@
 import React, { FC, useState } from 'react';
 import ExerciseList from '../components/ExerciseList/ExerciseList';
 import { MuscleGroupSelect } from '../components/MuscleGroupSelect/MuscleGroupSelect';
-import { ExerciseDifficulty, ExerciseMuscle } from '../components/types';
 import { DifficultyFilter } from '../components/DifficultyFilter/DifficultyFilter';
 import { Logo } from '../components/Logo/Logo';
+import useExerciseFilters from './hooks/use-exercise-filters/use-exercise-filters';
 
 const Home: FC = () => {
-	const [selectedMuscle, setSelectedMuscle] = useState<ExerciseMuscle | ''>('');
-	const [selectedDifficulties, setSelectedDifficulties] = useState<
-		ExerciseDifficulty[]
-	>(['beginner', 'intermediate', 'expert']);
+	const {
+		selectedMuscle,
+		setSelectedMuscle,
+		selectedDifficulties,
+		setSelectedDifficulties,
+	} = useExerciseFilters();
 
 	return (
-		<div className='flex flex-col justify-center items-center'>
+		<div className='flex flex-col items-center min-h-screen'>
 			<Logo />
 
 			<div className='flex flex-col gap-6 w-full max-w-4xl mt-20 px-8'>
